requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
httpx>=0.27,<1.0
cachetools>=5.3,<7.0
tenacity>=8.3,<10.0
python-dotenv>=1.0,<2.0
pytest>=8.2,<9.0
respx>=0.21,<1.0
